+++ src/lib/app/EarthViewerPixi.svelte
@@ -7,7 +7,7 @@
 		computeBlendedImagesCycleZIndex,
 	} from '$lib/app/blendedImagesCycle';
 	import {getPixiScene} from '$lib/app/pixi';
-	import InteractiveSurface from '$lib/app/InteractiveSurface.svelte';
+	import Surface from '$lib/app/Surface.svelte';
 	import type Camera from '$lib/app/Camera.svelte';
 
 	// TODO looks like we could use `Pixi.Prepare` to make initial rendering smoother:
@@ -215,7 +215,7 @@
 	};
 </script>
 
-<InteractiveSurface
+<Surface
 	width={$width}
 	height={$height}
 	scale={$scale}
