+++ src/routes/starlit-hammock/+page.svelte
@@ -8,7 +8,7 @@
 	import {spaceImages, type ImageMeta} from '$lib/app/images';
 	import {getClock} from '$lib/app/clock';
 	import ImageCreditsCaption from '$lib/app/ImageCreditsCaption.svelte';
-	import InteractiveSurface from '$lib/app/InteractiveSurface.svelte';
+	import Surface from '$lib/app/Surface.svelte';
 	import Panel from '$lib/app/Panel.svelte';
 	import AboutPortalPreview from '$lib/portals/about/Preview.svelte';
 	import PortalPreview from '$lib/app/PortalPreview.svelte';
@@ -232,7 +232,7 @@
 	</div>
 {:else}
 	<div class="interaction-surface-wrapper" style="width: {width}px; height: {height}px;">
-		<InteractiveSurface {width} {height} {scale} {zoomCamera} {moveCamera} />
+		<Surface {width} {height} {scale} {zoomCamera} {moveCamera} />
 	</div>
 {/if}
 <div class="hud idle-fade">
